Average expected hits per turn & helper updates
Rename and refactor hits-per-turn reporting to compute average expected hits per turn and update related helpers and queries.

Key changes:
- Rename HitsPerTurnService -> AverageHitsPerTurnService and adjust API/bean usage. Compute averageExpectedHitsPerTurn (totalExpectedHits / totalNumberOfTurns) and filter players by completed games only.
- Update command HitsPerTurn to call the new service and remove the old minimum-expected-hits option.
- Replace anomaly filtering heuristic with a fixed threshold (ignore games with expectedHits >= 200).
- Change result formatting: label updated to "Average Expected Hits Per Turn", show total expected hits / total turns, optionally include tier text, and append an overall average.
- Add PlayerEntityRepository.findAllWithUsersByCompletedGame() to query completed games.
- Change Helper.median signature to median(Collection<? extends Number>) -> double; validate empty input and compute median as double.
- Update callers: cast median results where needed (AverageTurnTimeService, MedianTurnTimeService) and initialize cached median sentinel in MedianTurnTimeService.
- Remove a previous anomaly filter from DiceLuckService (no longer ignores by expectedHits/turns ratio).
- Update LifeTimeRecordService to use the new AverageHitsPerTurnService method for summary output.

Several small renames and accumulator field adjustments were made to reflect total/average semantics and make percentile/tiering behavior clearer.

+++ src/main/java/ti4/spring/service/statistics/AverageTurnTimeService.java
@@ -74,7 +74,7 @@ public class AverageTurnTimeService {
             sb.append(DateTimeHelper.getTimeRepresentationToSeconds(stats.getAverage()));
 
             if (showMedian) {
-                long median = Helper.median(stats.gameAverages);
+                long median = (long) Helper.median(stats.gameAverages);
                 sb.append(" (median: ")
                         .append(DateTimeHelper.getTimeRepresentationToSeconds(median))
                         .append(")");
